+++ uae_src/scsitape.cpp
@@ -256,7 +256,7 @@ static void next_file (struct scsi_data_tape *tape)
 		tape->zf = zfile_fopen (path, _T("rb"), ZFD_NORMAL);
 		if (log_tapeemu) {
 			if (tape->zf)
-				write_log(_T("TAPEEMU: File '%s' Size %lld\n"), path, zfile_size(tape->zf));
+				write_log(_T("TAPEEMU: File '%s' Size %lld\n"), path, (long long)zfile_size(tape->zf));
 			else
 				write_log(_T("TAPEEMU: File '%s' could not be opened\n"), path);
 		}
@@ -285,12 +285,12 @@ static void next_file (struct scsi_data_tape *tape)
 			break;
 		}
 		if (tape->zf && log_tapeemu) {
-			write_log(_T("TAPEEMU DIR: File '%s' Size %lld\n"), zfile_getname(tape->zf), zfile_size(tape->zf));
+			write_log(_T("TAPEEMU DIR: File '%s' Size %lld\n"), zfile_getname(tape->zf), (long long)zfile_size(tape->zf));
 		}
 	} else {
 		tape->zf = zfile_readdir_archive_open (tape->zd, _T("rb"));
 		if (log_tapeemu && tape->zf)
-			write_log(_T("TAPEEMU ARC: File '%s' Size %lld\n"), zfile_getname(tape->zf), zfile_size(tape->zf));
+			write_log(_T("TAPEEMU ARC: File '%s' Size %lld\n"), zfile_getname(tape->zf), (long long)zfile_size(tape->zf));
 	}
 	if (tape->zf) {
 		tape->file_number++;
@@ -330,7 +330,7 @@ static int tape_read (struct scsi_data_tape *tape, uae_u8 *scsi_data, int len, b
 			got = (int)zfile_fread(scsi_data, 1, len, tape->zf);
 			uae_s64 pos = zfile_ftell(tape->zf);
 			if (log_tapeemu)
-				write_log(_T("TAPEEMU READ: Requested %d, read %d, pos %lld, %lld remaining.\n"), len, got, pos, zfile_size(tape->zf) - pos);
+				write_log(_T("TAPEEMU READ: Requested %d, read %d, pos %lld, %lld remaining.\n"), len, got, (long long)pos, (long long)(zfile_size(tape->zf) - pos));
 		} else {
 			got = 0;
 			if (len > 0) {
